feat: extract optimize runtime foundation
