+++ app/api/include/deliveryoptimizer/api/endpoints/health_endpoint.hpp
@@ -1,11 +1,22 @@
 #pragma once
 
+#include <functional>
+#include <memory>
+
+#include <json/json.h>
+
 namespace drogon {
 class HttpAppFramework;
 }
 
 namespace deliveryoptimizer::api {
 
-void RegisterHealthEndpoint(drogon::HttpAppFramework& app);
+class ObservabilityRegistry;
+
+using HealthExtensionProvider = std::function<void(Json::Value& checks, bool& overall_ready)>;
+
+void RegisterHealthEndpoint(drogon::HttpAppFramework& app,
+                            std::shared_ptr<const ObservabilityRegistry> observability = nullptr,
+                            HealthExtensionProvider extension = {});
 
 } // namespace deliveryoptimizer::api
